+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 ### Unreleased
 - Improved performance of `atob` and `btoa`, [#1503](https://github.com/zloirock/core-js/issues/1503), [#1464](https://github.com/zloirock/core-js/issues/1464)
-- [`Iterator.range`](https://github.com/tc39/proposal-iterator.range) updated following the actual spec version to throw a `RangeError` on `NaN` `start` / `end` / `step`
+- [`Iterator.range`](https://github.com/tc39/proposal-iterator.range) updated following the actual spec version
+  - Throw a `RangeError` on `NaN` `start` / `end` / `step`
+  - Allow `null` as `optionOrStep`
 - `Math.atanh` has become slightly more correct with small values
 - Fixed one more case (`Iterator.prototype.take`) of a V8 ~ Chromium < 126 [bug](https://issues.chromium.org/issues/336839115)
 - Forced replacement of `Iterator.{ concat, zip, zipKeyed }` in the pure version for ensuring proper wrapped `Iterator` instances as the result